refactor(modules): reorganize file structure into modules
