+++ src/buttons/index.js
@@ -1,0 +1,1 @@
+export { Button } from './Button';
