+++ src/icons/index.js
@@ -1,0 +1,5 @@
+import HelixUI from 'helix-ui';
+
+export const ICONS = Object.keys(HelixUI.ICONS);
+
+export * from './Icon';
