+++ src/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+/* Export helix-react definition */
+export * from './buttons';
+export * from './icons';
