--- src/index.mjs
@@ -1,8 +1,0 @@
-/* Export helix-react definition */
-import Button from './Button';
-import Icon from './Icon';
-
-export default {
-    Button,
-    Icon,
-};
